# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AUIB Academic Advisor API",
    "description": "Degree planning for AUIB students. Planning endpoints are stateless and store nothing.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/programs": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Published programs, one entry per program with its versions (F11.2, F0.4)",
        "operationId": "programs_api_v1_programs_get",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(major|minor)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProgramSummaryOut"
                  },
                  "title": "Response Programs Api V1 Programs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/programs/{program_id}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "A program's requirement tree",
        "operationId": "program_detail_api_v1_programs__program_id__get",
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Program Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/programs/{program_id}/insights": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Gateway courses and the longest prerequisite chain (F3.2)",
        "operationId": "insights_api_v1_programs__program_id__insights_get",
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Program Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/courses": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Search the catalog",
        "operationId": "courses_api_v1_courses_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 60
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "subject",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[A-Za-z]{2,4}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/courses/{code}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "A course with its rules shown beside their source text (F0.2)",
        "operationId": "course_detail_api_v1_courses__code__get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CourseListOut": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "courses": {
            "items": {
              "$ref": "#/components/schemas/CourseRef"
            },
            "type": "array",
            "title": "Courses"
          }
        },
        "type": "object",
        "required": [
          "total",
          "courses"
        ],
        "title": "CourseListOut"
      },
      "CourseOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "units": {
            "type": "number",
            "title": "Units"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "component": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Component"
          },
          "notices": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notices"
          },
          "offered_terms": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offered Terms",
            "description": "Seasons it runs in; null means every Fall and Spring"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/RuleOut"
            },
            "type": "array",
            "title": "Rules"
          },
          "unlocks": {
            "items": {
              "$ref": "#/components/schemas/CourseRef"
            },
            "type": "array",
            "title": "Unlocks"
          },
          "groups": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Groups"
          },
          "offerings": {
            "items": {
              "$ref": "#/components/schemas/TermOfferingsOut"
            },
            "type": "array",
            "title": "Offerings",
            "description": "Sections on published schedules, this term onward"
          }
        },
        "type": "object",
        "required": [
          "code",
          "title",
          "units",
          "description",
          "component",
          "notices",
          "offered_terms",
          "rules",
          "unlocks",
          "groups"
        ],
        "title": "CourseOut"
      },
      "CourseRef": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "units": {
            "type": "number",
            "title": "Units"
          }
        },
        "type": "object",
        "required": [
          "code",
          "title",
          "units"
        ],
        "title": "CourseRef"
      },
...
      "GatewayOut": {
        "properties": {
          "course": {
            "$ref": "#/components/schemas/CourseRef"
          },
          "dependents": {
            "type": "integer",
            "title": "Dependents"
          },
          "direct_dependents": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Direct Dependents"
          }
        },
        "type": "object",
        "required": [
          "course",
          "dependents",
          "direct_dependents"
        ],
        "title": "GatewayOut"
      },
...
      "GroupOut": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "units_required": {
            "type": "number",
            "title": "Units Required"
          },
          "requires_all": {
            "type": "boolean",
            "title": "Requires All"
          },
          "courses": {
            "items": {
              "$ref": "#/components/schemas/CourseRef"
            },
            "type": "array",
            "title": "Courses"
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/GroupOut"
            },
            "type": "array",
            "title": "Children"
          }
        },
        "type": "object",
        "required": [
          "key",
          "title",
          "label",
          "role",
          "units_required",
          "requires_all",
          "courses",
          "children"
        ],
        "title": "GroupOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InsightsOut": {
        "properties": {
          "gateways": {
            "items": {
              "$ref": "#/components/schemas/GatewayOut"
            },
            "type": "array",
            "title": "Gateways"
          },
          "longest_chain": {
            "items": {
              "$ref": "#/components/schemas/CourseRef"
            },
            "type": "array",
            "title": "Longest Chain"
          }
        },
        "type": "object",
        "required": [
          "gateways",
          "longest_chain"
        ],
        "title": "InsightsOut"
      },
...
      "OfferingOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "section": {
            "type": "string",
            "title": "Section"
          },
          "days": {
            "type": "string",
            "title": "Days"
          },
          "time": {
            "type": "string",
            "title": "Time"
          },
          "instructor": {
            "type": "string",
            "title": "Instructor"
          },
          "room": {
            "type": "string",
            "title": "Room"
          }
        },
        "type": "object",
        "required": [
          "code",
          "title",
          "section",
          "days",
          "time",
          "instructor",
          "room"
        ],
        "title": "OfferingOut"
      },
...
      "ProgramDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "catalog_year": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalog Year"
          },
          "total_units": {
            "type": "number",
            "title": "Total Units"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "source_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Date"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/ProgramVersionOut"
            },
            "type": "array",
            "title": "Versions",
            "description": "Every version of the program, oldest first (F0.4)"
          },
          "root": {
            "$ref": "#/components/schemas/GroupOut"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "kind",
          "catalog_year",
          "total_units",
          "source",
          "source_date",
          "root"
        ],
        "title": "ProgramDetailOut"
      },
...
      "ProgramSummaryOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "catalog_year": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalog Year"
          },
          "total_units": {
            "type": "number",
            "title": "Total Units"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "source_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Date"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/ProgramVersionOut"
            },
            "type": "array",
            "title": "Versions",
            "description": "Every version of the program, oldest first (F0.4)"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "kind",
          "catalog_year",
          "total_units",
          "source",
          "source_date"
        ],
        "title": "ProgramSummaryOut"
      },
      "ProgramVersionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "catalog_year": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalog Year"
          },
          "valid_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From",
            "description": "The first entry term it applies to; null means from the start"
          },
          "applies_to": {
            "type": "string",
            "title": "Applies To",
            "description": "Whom it applies to, e.g. \"students who joined from Fall 2027 on\""
          },
          "in_use": {
            "type": "boolean",
            "title": "In Use",
            "description": "The version this plan follows",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "catalog_year",
          "valid_from",
          "applies_to"
        ],
        "title": "ProgramVersionOut"
      },
...
      "RuleOut": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "english": {
            "type": "string",
            "title": "English"
          },
          "rule": {
            "type": "string",
            "title": "Rule"
          },
          "source_text": {
            "type": "string",
            "title": "Source Text"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "reviewed": {
            "type": "boolean",
            "title": "Reviewed"
          },
          "overridden": {
            "type": "boolean",
            "title": "Overridden"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "english",
          "rule",
          "source_text",
          "status",
          "reviewed",
          "overridden"
        ],
        "title": "RuleOut"
      },
...
      "TermOfferingsOut": {
        "properties": {
          "term": {
            "$ref": "#/components/schemas/TermOut"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/OfferingOut"
            },
            "type": "array",
            "title": "Sections"
          }
        },
        "type": "object",
        "required": [
          "term",
          "sections"
        ],
        "title": "TermOfferingsOut"
      },
      "TermOut": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "season": {
            "type": "string",
            "enum": [
              "Spring",
              "Summer",
              "Fall"
            ],
            "title": "Season"
          }
        },
        "type": "object",
        "required": [
          "label",
          "year",
          "season"
        ],
        "title": "TermOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Admin token (ADVISOR_ADMIN_TOKEN)",
        "scheme": "bearer"
      }
    }
  }
}